rtw89u: M3 association wiring (state/TX command isolation, peer snapshots)
Per-request state and TX command structs replace the shared
sc_cmd_state/sc_cmd_arg slots; a generation counter lets the worker
drop superseded or detach-orphaned commands.  Peer info (BSSID, AID,
rates, beacon interval, vif state) is snapshotted across the net80211
boundary so the chip path never retains net80211 node pointers.
Management TX dequeues ic_mgtq with its node reference again.  The
station lifecycle maps NOTEXIST->NONE->AUTH->ASSOC; AUTH/ASSOC program
the station and channel, RUN performs the full join, INIT/stop tear it
down, all under the wiphy mutex.

Board-verified on RTL8851BU (RK3568 ITX, RK3568_MFS #108):
- wpa_supplicant v2.11 -D bsd associates with the WPA2-PSK/CCMP AP;
  net80211 1->2->3->4, 4-way handshake completes, CTRL-EVENT-CONNECTED
- dhcpcd obtains 192.168.0.200/24 over the Wi-Fi link
- ping -c 20 gateway x2: 20/20 0.0% loss (avg 6.50/6.62 ms, 0 dup);
  ccmp crypto done in s/w 1914, CCMP MIC failures 0, deauth 0

+++ sys/external/bsd/rtw89/netbsd/rtw89_mac80211.h
@@ -1221,7 +1221,7 @@ struct ieee80211_sband_iftype_data {
 struct ieee80211_link_sta {
 	u8			link_id;
 	u8			addr[ETH_ALEN];
-	u8			supp_rates[IEEE80211_NUM_BANDS];
+	u32			supp_rates[IEEE80211_NUM_BANDS];
 	u8			rx_nss;
 	struct ieee80211_sta	*sta;
 	struct {
@@ -1249,7 +1249,7 @@ struct ieee80211_sta {
 	struct ieee80211_sta_ht_cap ht_cap;
 	struct ieee80211_sta_vht_cap vht_cap;
 	struct ieee80211_txq	*txq[IEEE80211_NUM_TIDS];
-	u8			supp_rates[IEEE80211_NUM_BANDS];
+	u32			supp_rates[IEEE80211_NUM_BANDS];
 	struct ieee80211_link_sta deflink;
 	/* holds struct_size_t(struct rtw89_sta, links_inst, 1): 304 + 568 */
 	u8			drv_priv[880] __aligned(sizeof(void *));
